Allow freestanding optional arguments
Sometimes we need optional arguments that can be provided on their own
without necessitating providing the normal positional arguments (e.g.
`--help` to print a help message and exit). This adds the `freestanding`
option to arguments, when a single freestanding argument is provided it
overrides the requirement that all positional arguments are provided.